+++ R/survival_reg-data.R
@@ -296,4 +296,154 @@ make_survival_reg_flexsurv <- function() {
   )
 }
 
+make_survival_reg_flexsurvspline <- function() {
+
+  parsnip::set_model_engine("survival_reg",
+                            mode = "censored regression",
+                            eng = "flexsurvspline")
+  parsnip::set_dependency("survival_reg",
+                          eng = "flexsurvspline",
+                          pkg = "flexsurv",
+                          mode = "censored regression")
+  parsnip::set_dependency("survival_reg",
+                          eng = "flexsurvspline",
+                          pkg = "survival",
+                          mode = "censored regression")
+  parsnip::set_dependency("survival_reg",
+                          eng = "flexsurvspline",
+                          pkg = "censored",
+                          mode = "censored regression")
+
+  parsnip::set_fit(
+    model = "survival_reg",
+    eng = "flexsurvspline",
+    mode = "censored regression",
+    value = list(
+      interface = "formula",
+      protect = c("formula", "data", "weights"),
+      func = c(pkg = "flexsurv", fun = "flexsurvspline"),
+      defaults = list()
+    )
+  )
+
+  parsnip::set_encoding(
+    model = "survival_reg",
+    eng = "flexsurvspline",
+    mode = "censored regression",
+    options = list(
+      predictor_indicators = "none",
+      compute_intercept = FALSE,
+      remove_intercept = FALSE,
+      allow_sparse_x = FALSE
+    )
+  )
+
+  parsnip::set_pred(
+    model = "survival_reg",
+    eng = "flexsurvspline",
+    mode = "censored regression",
+    type = "time",
+    value = list(
+      pre = NULL,
+      post = NULL,
+      func = c(fun = "predict"),
+      args =
+        list(
+          object = rlang::expr(object$fit),
+          newdata = rlang::expr(new_data),
+          type = "mean"
+        )
+    )
+  )
+
+  parsnip::set_pred(
+    model = "survival_reg",
+    eng = "flexsurvspline",
+    mode = "censored regression",
+    type = "quantile",
+    value = list(
+      pre = NULL,
+      post = NULL,
+      func = c(fun = "predict"),
+      args =
+        list(
+          object = rlang::expr(object$fit),
+          newdata = rlang::expr(new_data),
+          type = "quantile",
+          p = rlang::expr(quantile),
+          conf.int = rlang::expr(interval == "confidence"),
+          conf.level = rlang::expr(level)
+        )
+    )
+  )
+
+  parsnip::set_pred(
+    model = "survival_reg",
+    eng = "flexsurvspline",
+    mode = "censored regression",
+    type = "hazard",
+    value = list(
+      pre = NULL,
+      post = NULL,
+      func = c(fun = "predict"),
+      args =
+        list(
+          object = rlang::expr(object$fit),
+          newdata = rlang::expr(new_data),
+          type = "hazard",
+          times = rlang::expr(time)
+        )
+    )
+  )
+
+  parsnip::set_pred(
+    model = "survival_reg",
+    eng = "flexsurvspline",
+    mode = "censored regression",
+    type = "survival",
+    value = list(
+      pre = NULL,
+      post = NULL,
+      func = c(fun = "predict"),
+      args =
+        list(
+          object = expr(object$fit),
+          newdata = expr(new_data),
+          type = "survival",
+          times = expr(time),
+          conf.int = rlang::expr(interval == "confidence"),
+          conf.level = rlang::expr(level)
+        )
+    )
+  )
+
+  parsnip::set_pred(
+    model = "survival_reg",
+    eng = "flexsurvspline",
+    mode = "censored regression",
+    type = "linear_pred",
+    value = list(
+      pre = NULL,
+      post = function(results, object) {
+        # flexsurv returns on the natural scale of the location parameter
+        # thus transform to the unrestricted scale before returning
+        location_name <- object$fit$dlist$location
+        location_index <- which(object$fit$dlist$pars == location_name)
+        transformation <- object$fit$dlist$transforms[[location_index]]
+
+        results %>%
+          dplyr::mutate(.pred_linear_pred = transformation(.pred_link)) %>%
+          dplyr::select(.pred_linear_pred)
+      },
+      func = c(fun = "predict"),
+      args =
+        list(
+          object = expr(object$fit),
+          newdata = expr(new_data),
+          type = "linear"
+        )
+    )
+  )
+}
+
 # nocov end