+++ tests/testthat/helper-case-weights.R
@@ -1,3 +1,10 @@
+expect_unequal <-   function(object,
+                             expected,
+                             ...,
+                             tolerance = if (edition_get() >= 3) testthat_tolerance()) {
+  expect_true(!compare(object, expected, tolerance = tolerance, ...)$equal)
+}
+
 make_cens_wts <- function() {
   data(time_to_million, package = "censored", envir = rlang::current_env())