+++ NEWS.md
@@ -2,7 +2,9 @@
 
 * Added the new `"aorsf"` engine for `rand_forest()` for accelerated oblique random survival forests with the aorsf package (@bcjaeger, #211).
 
-* The matrix interface for fitting `fit_xy()` now works for censored regression models, with the exception of `decision_tree(engine = "rpart")` and `survival_reg(engine = "flexsurv")` (#225).
+* Added the new `flexsurvspline` engine for `survival_reg()` (@mattwarkentin, #213).
+
+* The matrix interface for fitting `fit_xy()` now works for censored regression models, with the exception of `decision_tree(engine = "rpart")` and `survival_reg()` with engines `"flexsurv"` and `"flexsurvspline"` (#225).
 
 * Predictions of survival probabilities are now calculated via `summary.survfit()` for `proportional_hazards()` models with the `"survival"` and `"glmnet"` engines, `bag_tree()` models with the `"rpart"` engine, `decision_tree()` models with the `"partykit"` engines, as well as `rand_forest()` models with the `"partykit"` engine (#221, #224).